+++ src/react/index.ts
@@ -1,7 +1,8 @@
 /// <reference types="react" />
-import type { Derived } from "@tanstack/store";
+
 import { deepEqual, shallowEqual } from "fast-equals";
 import { useRef, useSyncExternalStore } from "react";
+
 import type { NusmStore } from "../types";
 
 export type NoInfer<T> = [T][T extends unknown ? 0 : never];
@@ -15,25 +16,26 @@ export function useStore<TState, TSelected = NoInfer<TState>>(
   selector?: (state: NoInfer<TState>) => TSelected,
   options?: UseStoreOptions,
 ): TSelected;
-export function useStore<TState, TSelected = NoInfer<TState>>(
-  store: Derived<TState>,
-  selector?: (state: NoInfer<TState>) => TSelected,
-  options?: UseStoreOptions,
-): TSelected;
+
 export function useStore<TState, TSelected = NoInfer<TState>>(
   store: NusmStore<TState>,
   selector?: (state: NoInfer<TState>) => TSelected,
   options?: UseStoreOptions,
 ): TSelected;
 export function useStore<TState, TSelected = NoInfer<TState>>(
-  store: NusmStore<TState> | Derived<TState>,
+  store: NusmStore<TState>,
   selector: (state: NoInfer<TState>) => TSelected = (d) => d as TSelected,
   options: UseStoreOptions = {},
 ): TSelected {
   const equal = options.equal ? deepEqual : shallowEqual;
   const getSnapshot = () => store.state;
   const snapshot = useSyncExternalStore(
-    store.subscribe,
+    (notify) => {
+      const subscription = store.subscribe(notify);
+      return typeof subscription === "function"
+        ? subscription
+        : subscription.unsubscribe;
+    },
     getSnapshot,
     getSnapshot,
   );